Kept the meters' decimal digit through the analytics reading path
Bulk flow meters carry a decimal digit — FlowVision extracts it as the red last digit — and tenant_<state>.flow_reading_table records it as NUMERIC. Analytics narrowed both readings to INT, so up to 0.5 m3 was discarded per reading and a daily volume, being the difference of two readings, could be off by up to 1000 L. That error was invisible while the column held cubic metres and became visible once V45 denominated it in litres.

V47 widens fact_meter_reading_table.extracted_reading and .confirmed_reading to bare NUMERIC — the source's own type, not a scale the warehouse invents — and the readings are BigDecimal from the event through the entity to the calculator. The delta is now subtracted at the readings' own precision and rounded once, on the way into litres; rounding each reading first and subtracting after is what cost the 1000 L.

water_quantity stays BIGINT litres. A litre is three orders of magnitude finer than the meters resolve, so rounding there discards nothing real and every consumer — the Long DTOs, the total_water_supplied_liters CSV column, the frontend — is untouched. The rounding is HALF_UP because the shared recompute definition expresses the same conversion as Postgres ROUND(), which is half-away-from-zero; on the non-negative values GREATEST admits the two coincide, and WaterQuantityBackfillParityIntegrationTest now asserts that on a delta landing exactly on a half litre.

One deliberate consequence: the compliant/anomalous submission counts compare extracted_reading to confirmed_reading, and that comparison becomes exact. An operator correcting 1235.4 to 1235.6 was previously counted as "the AI was right"; it now counts as a correction. Expect a step in that metric at the deploy, not a regression.

No history is repaired and none can be — the discarded decimals were lost at publish time, and the full-precision flow_reading_table lives in a different production database. Readings written from here on carry their decimals; earlier ones stay whole.

Deploy this before telemetry-service stops rounding  at the publisher: a whole-number reading off the topic still deserialises into the widened fields unchanged, so this service is safe to run ahead of that change.

+++ backend/analytics-service/src/main/java/org/arghyam/jalsoochak/analytics/service/water/WaterQuantityContext.java
@@ -2,6 +2,7 @@ package org.arghyam.jalsoochak.analytics.service.water;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 /**
@@ -21,14 +22,18 @@ import java.time.LocalDate;
  * Because interpretation is delegated to the channel's calculator, a reading is only ever
  * processed by the calculator that understands its units; see
  * {@link WaterQuantityCalculatorRegistry#resolve(Integer)}.
+ *
+ * <p>The readings are {@code BigDecimal} because they are decimal at the source — bulk flow meters
+ * carry a decimal digit — and because a channel whose reading is not a whole count of anything
+ * (ELM's kWh, PDU's pump minutes) cannot be represented otherwise.
  */
 @Builder
 public record WaterQuantityContext(
         Integer tenantId,
         Integer schemeId,
         LocalDate readingDate,
-        Integer currentReading,
-        Integer previousReading,
+        BigDecimal currentReading,
+        BigDecimal previousReading,
         Integer channel
 ) {
 }
